refactor: Change feature folder files to ts
