+++ src/feature/Footer/constant.ts
@@ -1,5 +1,6 @@
-export const TEXT = {
+const TEXT = {
   codeit: "©codeit - 2023",
   privacyPolicy: "Privacy Policy",
   faq: "FAQ",
 };
+export default TEXT;
