--- src/feature/Layout/index.js
@@ -1,1 +1,0 @@
-export * from "./Layout";
