--- src/feature/NavigationBar/index.js
@@ -1,1 +1,0 @@
-export * from "./NavigationBar";
