+++ src/ui/Cta/Cta.tsx
@@ -1,5 +1,7 @@
 import "./Cta.css";
 
-export const Cta = ({ children }) => {
+const Cta = ({ children }: any) => {
   return <div className="Cta">{children}</div>;
 };
+
+export default Cta;
