--- src/ui/Cta/index.js
@@ -1,1 +1,0 @@
-export * from "./Cta";
